Revert "Re-Fixing bugs (Sorry)"
This reverts commit 00a9a4a7caeaeb3316e9ae39f5b706b59b9b3ed3.

+++ source/core/src/main/com/deco2800/game/GdxGame.java
@@ -121,7 +121,7 @@ public class GdxGame extends Game {
   }
 
   private int getCompletionTime() {
-    return Math.round((float)timeScore/1000);
+    return Math.round(timeScore/1000);
   }
 
   public enum ScreenType {
